Fix some small bugs
https://www.youtube.com/watch?v=uQxF3HNQBp8

The bug where there is a typo in the title screen was already fixed in
the code, apparently I just forgot to recompile the jar. Actually, I
might remember that from 2014 when I showed the game to my dad.

Then there was a report of the numpad keys not working. This should be
fixed by adding some more KeyEvent enum values to the switch
statement (fingers crossed).

Also, since this was a super old project that I had done in Eclipse, I
just cleaned up everything and added a Makefile that gets the job
done. So now it's reasonably easy to get the thing compiling again. I
compiled with JDK 1.8 because that's what I have lying around
currently for Minecraft. No idea if it would behave differently in
newer versions.